Implement AI_FLAG_KNOWLEDGABLE
AI emulates basic knowledge of Pokémon's moves and abilities by reading them outright. This would be cheating if it was a human player, but for the AI this simply lets it act as if it predicted the player to have certain moves. Because this is an AI flag only some trainers will have this ability.