Use Fly/Flash from party menu if Pokémon is compatible
This commit allows Fly and/or Flash to be used from the party menu if the Pokémon is able to learn (and use) the relevant HMs, without them needing to be taught. If the Pokémon already knows 4 HM moves, no moves will be added to the action list. If the Pokémon knows less than 4 HM moves, Fly will be prioritized over Flash.

The necessary badges are still required to use these moves in the field.